python-sdet-backend/requirements.txt
requests
mysql-connector-python
behave